test: Create custom mockito verifiers for Hibernate entities. (#250)
test: Create custom mockito verifiers for Hibernate entities.

Previously, due to the requirements of Hibernate to be independent of comparison based on entity fields using equals, previous implementation used getClass and id field for entities equality. That made it impossible to compare expected and actual results in tests as well as proper mocking. Previous temporary solution used captor to intercept saved entity and compare based on the entity fields. This commit solves this problem by creating custom Argument Matchers for most used Hibernate entities.

- Create custom Argument Matchers for AuthoritySourceFile and Authority classes.
- Replace previous assertions with Matchers in verifiers.

Closes MODELINKS-209