fix(authority-prefix): delete sequence during deletion of authority-source-file.
Previously, during creation of the authority-source-file with the same prefix as previously deleted authority-source-file, program generated an error that states when you try to recreate a sequence with the same name after deleting the associated row, it encounters a conflict because it thinks the sequence already exists. This commit solves this error by deleting sequence linked to authority-source-file during its deletion.

- add deleteSequence method
- refactor deleteAuthoritySourceFile_positive_deleteExistingEntity test

Closes MODELINKS-211